Changed all float types to double types. Added more point operators. Added unit tests to fully cover Point and Rectangle types. Integrated Biological and California pre-generated datasets into benchmark. Unified statistics gathering and reporting across R/R+/NIR+ trees. Fixed NIR node splitting.

+++ src/include/nirtree/node.h
@@ -52,7 +52,7 @@ namespace nirtree
 			struct Partition
 			{
 				unsigned dimension;
-				float location;
+				double location;
 			};
 
 			Node *parent;
